Decouples SiteSearch from Navigation
