+++ app/components/ui/SiteSearch.module.scss
@@ -1,0 +1,24 @@
+@import "../../styles/utils/_helpers.scss";
+
+.navSearch {
+  display: inline-block;
+
+  .searchField {
+    $font-size: calc-em(15px);
+    background-repeat: no-repeat;
+    background-image: url(../../assets/img/ic-search.png);
+    background-size: 18px 18px;
+    background-position: calc-em(15px) center;
+    font-size: $font-size;
+    padding-left: calc-em(45px);
+    width: calc-em(400px);
+    height: calc-em(40px);
+    border: 0;
+    background-color: $color-grey1;
+
+    &::placeholder {
+      font-size: $font-size;
+      font-style: normal;
+    }
+  }
+}
